+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 4.0.0
+* Requires Laravel v7.7+
+* Requires php v7.3+
+* `BandwidthClient` class has been remove in favor of Laravel inbuilt [HTTP client](https://laravel.com/docs/7.x/http-client)
+
 ## 3.0.0
 * Requires Laravel v6.x+
 * Test against php v7.4